+++ iqm_pulse/src/iqm/pulse/gates/default_gates.py
@@ -34,7 +34,7 @@ from iqm.pulse.gates.cz import (
 )
 from iqm.pulse.gates.delay import Delay
 from iqm.pulse.gates.flux_multiplexer import FluxMultiplexer_SampleLinear
-from iqm.pulse.gates.measure import Measure_Constant
+from iqm.pulse.gates.measure import Measure_Constant, Shelved_Measure_Constant
 from iqm.pulse.gates.move import MOVE_CRF_CRF, MOVE_TGSS_CRF
 from iqm.pulse.gates.prx import (
     PRX_DRAGCosineRiseFall,
@@ -60,7 +60,11 @@ for different gates."""
 _default_implementations = {
     "barrier": {"": Barrier},
     "delay": {"wait": Delay},
-    "measure": {"constant": Measure_Constant, "constant_qnd": Measure_Constant},
+    "measure": {
+        "constant": Measure_Constant,
+        "constant_qnd": Measure_Constant,
+        "shelved_constant": Shelved_Measure_Constant,
+    },
     "prx": {
         "drag_gaussian": PRX_DRAGGaussian,
         "drag_crf": PRX_DRAGCosineRiseFall,